Organize project structure and enhance maintainability
