+++ README.md
@@ -1,2 +1,47 @@
-# MC-Mod-Updater
-A simple yet effective Minecraft mod updater made in python 3.11.8.
+# MC Mod Updater
+
+MC Mod Updater is a simple yet effective Python application designed to facilitate the updating and maintenance of Minecraft mods. It allows users to effortlessly update their installed mods and remove any unused ones.
+
+## Features
+
+- **Mod Updating**: Automatically checks for updates to installed mods and downloads the latest versions from a specified GitHub repository.
+- **Mod Removal**: Easily remove any mods that are no longer needed or in use.
+- **Graphical User Interface (GUI)**: Provides a user-friendly interface for initiating mod updates and removals.
+
+## Requirements
+
+- Python 3.11.8 or later
+- Dependencies listed in `requirements.txt`
+
+## Installation
+
+1. Clone the repository:
+
+    ```python
+    git clone https://github.com/yourusername/MC-Mod-Updater.git
+    ```
+
+2. Install dependencies:
+
+    ```python
+    pip install -r requirements.txt
+    ```
+
+## Usage
+Run the application:
+
+    ```python
+    python main.py
+    ```
+
+## Configuration
+
+* Repository URL: The GitHub repository URL where the mod files are hosted.
+* Mods Folder: The local folder where mods are stored.
+
+## Contributing
+
+Contributions are welcome! If you'd like to contribute to this project, please fork the repository and submit a pull request.
+
+## License
+This project is licensed under the GPL-3.0 License - see the [LICENSE](https://raw.githubusercontent.com/MotanOfficial/MC-Mod-Updater/main/LICENSE) file for details.
