+++ main.py
@@ -1,0 +1,4 @@
+from src.gui import run
+
+if __name__ == "__main__":
+    run()
