+++ src/__init__.py
@@ -1,0 +1,14 @@
+REPO_URL = 'https://api.github.com/repos/MotanOfficial/MC-Mod-Updater/contents/Mods'
+MODS_FOLDER = 'mods'
+
+__all__ = ['REPO_URL', 'MODS_FOLDER']
+
+__version__ = '1.0.0'
+__author__ = ['MotanOfficial', 'Cathie02']
+__github__ = ['https://github.com/MotanOfficial', 'https://github.com/Cathie02']
+
+__doc__ = """
+MC Mod Updater
+
+This package provides a simple GUI application to update Minecraft mods from a GitHub repository.
+"""
